Add ensure that address points to contract

+++ packages/emissions-dashboard/pages/index.tsx
@@ -2,6 +2,9 @@ import { NavBar } from '@popcorn/ui/components/popcorn/emissions-dashboard/NavBa
 import { useRouter } from 'next/router';
 import React, { useEffect, useState } from 'react';
 import web3 from 'web3';
+import { useWeb3React } from '@web3-react/core';
+import { Web3Provider } from '@ethersproject/providers';
+import { connectors } from '../context/Web3/connectors';
 
 const user = {
   name: 'Tom Cook',
@@ -22,6 +25,12 @@ const IndexPage = () => {
   const router = useRouter();
   const [open, setOpen] = useState<boolean>(false);
   const [errorMessage, setErrorMessage] = useState<string>('');
+  const context = useWeb3React<Web3Provider>();
+  const {
+    library,
+    activate,
+    active,
+  } = context;
 
   useEffect(() => {
     if (typeof window !== 'undefined' && window.location.pathname !== '/') {
@@ -29,24 +38,40 @@ const IndexPage = () => {
     }
   }, [router.pathname]);
 
-  const addContract = (contractAddress: string): void => {
+  useEffect(() => {
+    if (!active) {
+      activate(connectors.Network);
+    }
+  }, [active]);
+
+  const addContract = async(contractAddress: string): Promise<void> => {
+    const enterMessage: string =  "Please enter a valid address";
+    let message: string;
     if(contractAddress){
       if(web3.utils.isAddress(contractAddress)){
-        if (localStorage.getItem('contracts')) {
-          const existingContracts = JSON.parse(localStorage.getItem('contracts'));
-          existingContracts.push(contractAddress);
-          localStorage.setItem('contracts', JSON.stringify(existingContracts));
+        const code = await library.getCode(contractAddress);
+        const isConnected = !(code === "0x0" || code === "0x")
+        if(isConnected){
+          if (localStorage.getItem('contracts')) {
+            const existingContracts = JSON.parse(localStorage.getItem('contracts'));
+            if (!existingContracts.includes(contractAddress)) {
+              existingContracts.push(contractAddress);
+              localStorage.setItem('contracts', JSON.stringify(existingContracts));
+            }
+          } else {
+            localStorage.setItem('contracts', JSON.stringify([contractAddress]));
+          }
         } else {
-          localStorage.setItem('contracts', JSON.stringify([contractAddress]));
+          message = `The address you entered does not point to a valid Ethereum contract. ${enterMessage}`
         }
-        setOpen(false);
       } else {
-        const message = "The address you entered is not a valid Ethereum contract. Please enter a valid address."
-        setErrorMessage(message);
+        message = `The address you entered is not a valid Ethereum contract. ${enterMessage}`
       }
     } else {
-      setErrorMessage("No Contract Address was provided. Please enter a valid address");
+      message = `No Contract Address was provided. ${enterMessage}`;
     }
+    setErrorMessage(message);
+    setOpen(false);
   };
 
   const openAddContractModal = ():void => {
